create `add_txn` fxn to pool
