+++ Cargo.toml
@@ -10,6 +10,7 @@ anyhow = "1.0.65"
 blake3 = { version = "1.3.1", features = ["digest", "traits-preview"] }
 chrono = "0.4.22"
 ed25519 = "1.5.2"
+failure = "0.1.8"
 openssl = "0.10.42"
 p256 = { version = "0.11.1", features = ["ecdsa-core"] }
 rand = { version = "0.8.5", features = ["std_rng"] }
